Trait mpi_futures::buffer::OwnedBufferMut

pub trait OwnedBufferMut {
    type BufferMut: BufferMut + ?Sized;
    type Anchor;
    unsafe fn into_buffer_mut<'a>(self)
                                  -> (Self::Anchor, &'a mut Self::BufferMut);
}

An owned buffer that can be modified.

Associated Types

type BufferMut: BufferMut + ?Sized

type Anchor

The anchor is a (possibly opaque) object that is responsible for performing clean up after the buffer is no longer needed.

To avoid aliasing issues, the anchor is usually represented by a separate data type that contains raw pointers (rather than an owned data type like Vec, which internally contains a Unique).

Required Methods

unsafe fn into_buffer_mut<'a>(self) -> (Self::Anchor, &'a mut Self::BufferMut)

Split the owned buffer into an anchor and a mutable buffer.

This function is unsafe because lifetime of the buffer is misleading: it will only last as long as the anchor survives.