Trait mpi_futures::codec::Codec

pub trait Codec<'a> {
    type Message;
    type FutureMessage: Future<Item = Self::Message, Error = Void>;
    fn decode<R: RecvInto<'a>>(&self,
                               status: Status,
                               r: R)
                               -> (R::Output, Self::FutureMessage);
    fn encode<S: SendFrom<'a>>(&self, msg: Self::Message, s: S) -> S::Output;
}

Associated Types

type Message

Type of each message produced by Incoming and consumed by send.

(Not to be confused with MPI_Message, which is more of a handle to a pending message.)

type FutureMessage: Future<Item = Self::Message, Error = Void>

Required Methods

fn decode<R: RecvInto<'a>>(&self,
status: Status,
r: R)
-> (R::Output, Self::FutureMessage)

fn encode<S: SendFrom<'a>>(&self, msg: Self::Message, s: S) -> S::Output