Struct mpi_futures::request_poll::RequestPoll

pub struct RequestPoll<'a> { /* fields omitted */ }

Manages a collection of requests and keeps their associated buffers alive.

When RequestPoll is dropped, all pending requests will be canceled when possible and waited on.

Methods

impl<'a> RequestPoll<'a>

fn test(&mut self)

Non-blocking test to see if some of the requests have completed. For any request that is complete, the corresponding callback is called.

fn wait(&mut self)

Block until at least one request has completed. Otherwise functions similar to test.

fn mrecv<B, F>(&mut self, msg: Message, buf: B, callback: F) where B: OwnedBufferMut, B::Anchor: 'a, F: FnOnce(B::Anchor) + 'a

Perform a matched receive on a message.

fn send<D, B, F>(&mut self, dest: D, buf: B, tag: u16, callback: F) where D: Destination, B: OwnedBuffer + 'a, F: FnOnce(B) + 'a

Send a message.

unsafe fn insert<F>(&mut self,
request: MPI_Request,
callback: F,
cancelable: bool) where F: FnOnce() + 'a

Insert a request to be monitored.

cancelable indicates whether MPI_Cancel will work on the request (true for receiving requests, false for all other requests).

Unsafety

The request must be valid (in particular, not MPI_REQUEST_NULL). The buffers associated with the request must survive so long as the callback remains alive.

fn reserve_one(&mut self)

Allocate room for a single request if necessary.

Panics

Panics if the internal capacity overflows usize.