Struct mpi_futures::switch::LinkedCodec

pub struct LinkedCodec<'a, C: Codec<'a> + Clone> {
    pub link: Link<'a>,
    pub codec: C,
}

A combined Link and Codec.

Fields
link: Link<'a>

The Link to the Switch.

codec: C

The message Codec.

Methods

impl<'a, C: Codec<'a> + Clone> LinkedCodec<'a, C>

fn incoming<S: Source>(&self, source: S) -> Incoming<'a, C, S>

Obtain a Stream of future incoming messages from the given source. Each message is decoded using the given codec.

fn incoming(&self, Source) -> Stream<Future<Codec::Message>>;

The stream will keep running until the Switch is closed, but you can stop the Stream at any time if you aren't expecting to receive messages. You can even create a new incoming stream every time you want to receive a message.

Just try to avoid running multiple overlapping incoming streams simultaneously, as that could cause messages to be split between the streams in a non-deterministic manner.

fn send<D: Destination>(&self, dest: D, msg: C::Message) -> Send<'a, C, D>

Send a message asynchronously, returning a Future that completes when the send does.

fn send(&self, Destination, Codec::Message) -> Future<()>;

fn close(&self)

Same as Link::close.